refactor(handlers): split notification.rs god-file by domain
src/handlers/notification.rs was 2877 LOC: a dispatcher (handle_notification_impl)
routing to ~18 independent free-fn handlers, each a distinct notification domain
with no coupling between them, plus ~1200 LOC of tests.

Convert it to a module directory. NotificationHandler, its StanzaHandler impl, the
handle_notification_impl dispatcher, and the test module stay in mod.rs; the
handlers move into four cohesive submodules:
- device.rs: encrypt, account_sync (+ devices/identity/prekey/digest)
- groups.rs: server_sync, group, newsletter, mex, disappearing_mode
- privacy_business.rs: privacy_token, business
- profile.rs: picture, status, contacts (+ notification_timestamp/learn_contact_modify_mappings,
  whose only callers live in this group)

Pure intra-crate code-move: same codegen, no new dyn/Box/Arc/trait, no public API
change. Handlers become pub(crate); mod.rs glob-imports the submodules so the
dispatcher's calls are unchanged, and re-exports device::* (pub(crate)) so the
external crate::handlers::notification::handle_local_identity_change path
(device_registry.rs) keeps resolving. AccountSyncDevice's fields are pub(crate) so
the test module (now a parent) can still read them.

Verified: build, clippy --all-targets -D warnings, fmt, and the full whatsapp-rust
lib test suite (737 passing, incl. all 40 notification tests).
